# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codex3 API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/assets:
    get:
      summary: List assets
      parameters:
        - name: group_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Assets
...
  /api/v1/assets/{id}:
    delete:
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted asset
...